% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPeptides.R
\name{processPeptides}
\alias{processPeptides}
\title{Process and summarize MaxQuant peptide data}
\usage{
processPeptides(peptides_file, intensity_columns_file, protein_mapping_file)
}
\arguments{
\item{peptides_file}{Character. Path to the MaxQuant peptide data file.}

\item{intensity_columns_file}{Character. Path to the CSV file describing intensity columns and other metadata.}

\item{protein_mapping_file}{Character. Path to the CSV file containing protein mapping information.}
}
\value{
A named list with the following elements:
\itemize{
\item \code{dt.peptides}:  Filtered wide-format \code{data.table} of peptides with basic columns, intensity measurements, \code{Protein.name}, and \code{Protein.group}.
\item \code{dt.peptides.mean}: Data.table with \code{Mean.Intensity} per group across replicates.
\item \code{dt.peptides.mean.reps}: Data.table with \code{Intensity} per replicate per group.
\item \code{dt.peptides.count}: Counts of peptides per group in \code{dt.peptides.mean}.
\item \code{dt.peptides.count.reps}: Counts of peptides per replicate per group in \code{dt.peptides.mean.reps}.
\item \code{dt.int_col}: The intensity column metadata read from \code{intensity_columns_file}.
\item \code{grp_cols}: Character vector of grouping column names used.
\item \code{peptides_select_col.basic}: Character vector of basic peptide columns selected initially.
}
}
\description{
Read peptide information from MaxQuant output file, intensity column meatadata, and protein mapping;
filters out contaminants; computes per-replicate and per-group mean intensities;
and counts peptides per replicate and per group .
}
\examples{
\dontrun{
result <- processPeptides(
  peptides_file          = "../Data/peptides.txt",
  intensity_columns_file = "../Data/Intensity_columns.csv",
  protein_mapping_file   = "../Data/protein_mapping.csv"
)
dt.peptides      <- result$dt.peptides
dt.peptides.mean <- result$dt.peptides.mean
}
}
