% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPeptides.R
\name{processPeptides}
\alias{processPeptides}
\title{Process and summarize MaxQuant peptide data}
\usage{
processPeptides(peptides_file, intensity_columns_file, protein_mapping_file)
}
\arguments{
\item{peptides_file}{A \code{character} specifying the path to the MaxQuant peptide data file,
or a \code{data.frame}/\code{data.table} already loaded in R.}

\item{intensity_columns_file}{A \code{character} specifying the path to the CSV of intensity-column metadata,
or a \code{data.frame}/\code{data.table} already loaded in R.}

\item{protein_mapping_file}{A \code{character} specifying the path to the CSV of protein mapping information,
or a \code{data.frame}/\code{data.table} already loaded in R.}
}
\value{
A named list with the following elements:
\itemize{
\item \code{dt.peptides}:  Filtered wide-format \code{data.table} of peptides with basic columns, intensity measurements, \code{Protein.name}, and \code{Protein.group}.
\item \code{dt.peptides.int}: Data.table with \code{Mean.Intensity} per group across replicates.
\item \code{dt.peptides.int.reps}: Data.table with \code{Intensity} per replicate per group.
\item \code{dt.peptides.int.ttest}: Data.table for potential t.test analysis. Zero values are not dropped.
\item \code{dt.peptides.count}: Counts of peptides per group in \code{dt.peptides.int}.
\item \code{dt.peptides.count.reps}: Counts of peptides per replicate per group in \code{dt.peptides.int.reps}.
\item \code{dt.int_col}: The intensity column metadata read from \code{intensity_columns_file}.
\item \code{grp_cols}: Character vector of grouping column names used.
\item \code{peptides_select_col.basic}: Character vector of basic peptide columns selected initially.
}
}
\description{
Read peptide information from MaxQuant output file, intensity column meatadata, and protein mapping;
filters out contaminants; computes per-replicate and per-group mean intensities;
and counts peptides per replicate and per group .
Map protein name and protein group, calculate GRAVY score.
}
\examples{
\dontrun{
# 1) Reading from files
result <- processPeptides(
  peptides_file          = "../Data/peptides.txt",
  intensity_columns_file = "../Data/Intensity_columns.csv",
  protein_mapping_file   = "../Data/protein_mapping.csv"
)

# 2) Reading from in-memory files
library(data.table)
dt_pep <- fread("../Data/peptides.txt", integer64 = "double")
dt_lfq <- fread("../Data/Intensity_columns.csv")
dt_map <- fread("../Data/protein_mapping.csv")
result_2 <- processPeptides(
  peptides_file          = dt_pep,
  intensity_columns_file = dt_lfq,
  protein_mapping_file   = dt_map
)
}
}
